pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "signalscribe"
version = "0.7.0"
description = "A tool for transcribing and processing audio signals"
readme = "README.md"
authors = [{ name = "Signals Rising" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["audio", "transcription", "signal-processing"]
dependencies = [
    "watchdog",
    "rich",
    "pyyaml",
    "pywhispercpp",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = ["black", "pytest"]

[project.urls]
Homepage = "https://github.com/SignalsRising/SignalScribe"
"Bug Tracker" = "https://github.com/SignalsRising/SignalScribe/issues"

[project.scripts]
signalscribe = "SignalScribe.SignalScribe:main"

[tool.setuptools]
packages = ["SignalScribe"] 